+++ packages/builtin_plugins/mflowy/builtin_plugins/model/xgboost.py
@@ -6,11 +6,11 @@ from typing import Annotated
 
 from mflowy.builtin_plugins.cross_validation.types import DatasetLoader
 from mflowy.builtin_plugins.middlewares import inject_dataset_loader, inject_x_preprocessors
+from mflowy.builtin_plugins.model.study import ContinuousSpace, DiscreteSpace
 from mflowy.builtin_plugins.params_phaser import annotated_params_phaser
 from mflowy.driver.handler import handler
 from mflowy.utils.constants import RANDOM_STATE
 from mflowy.utils.logging import is_verbose
-from mflowy.utils.study import ContinuousSpace, DiscreteSpace
 from sklearn.compose import ColumnTransformer
 
 from ._pipeline import optimize, training